+++ src/renderer/config-blocks/EasyColor.svelte
@@ -30,252 +30,10 @@
 </script>
 
 <script lang="ts">
-  import { onMount, createEventDispatcher } from "svelte";
-  import { MeltCombo, MeltSelect } from "@intechstudio/grid-uikit";
-  import { GridScript } from "@intechstudio/grid-protocol";
-  import Toggle from "../main/user-interface/Toggle.svelte";
-  import { ElementType } from "@intechstudio/grid-protocol";
-  import SendFeedback from "../main/user-interface/SendFeedback.svelte";
-  import { Validator } from "./_validators";
-  import { Script } from "./_script_parsers.js";
-  import { LocalDefinitions } from "../runtime/runtime.store";
-  import {
-    ActionData,
-    GridAction,
-    GridElement,
-    GridEvent,
-  } from "./../runtime/runtime";
-  import { Grid } from "../lib/_utils.js";
-  import SliderColorPicker from "../main/user-interface/SliderColorPicker.svelte";
-  import SquareColorPicker from "../main/user-interface/SquareColorPicker.svelte";
-  import CircleColorPicker from "../main/user-interface/CircleColorPicker.svelte";
-  import RandomColorGenerator from "../main/user-interface/RandomColorGenerator.svelte";
-
+  import { GridAction } from "./../runtime/runtime";
+  import Temp from "./components/Temp.svelte";
   export let config: GridAction;
   export let index;
-
-  let event = config.parent as GridEvent;
-  const dispatch = createEventDispatcher();
-
-  const parameterNames = ["LED Number", "Layer", "Red", "Green", "Blue"];
-  const validators = [
-    (e) => {
-      return new Validator(e).NotEmpty().Result();
-    },
-    (e) => {
-      return new Validator(e).NotEmpty().Result();
-    },
-    (e) => {
-      return new Validator(e).NotEmpty().Result();
-    },
-    (e) => {
-      return new Validator(e).NotEmpty().Result();
-    },
-    (e) => {
-      return new Validator(e).NotEmpty().Result();
-    },
-  ];
-
-  let scriptSegments = [];
-
-  let beautyMode = 0;
-  let beautify = true;
-
-  const defaultColor = new Grid.RGB(255, 0, 0).toHSL();
-  let color: Grid.HSL = defaultColor;
-
-  $: handleConfigChange($config);
-
-  function handleConfigChange(config: ActionData) {
-    return;
-    const _segments = Script.toSegments({
-      short: config.short,
-      script: config.script,
-    });
-
-    // handle legacy and new beautify command
-    if (_segments.length == 6) {
-      beautify = +_segments[5] == 1 ? false : true; // check on number instead of string!
-      scriptSegments = _segments.slice(0, -1);
-    } else {
-      scriptSegments = _segments;
-      beautify = true;
-    }
-
-    color =
-      Grid.parseRGB(
-        scriptSegments[2],
-        scriptSegments[3],
-        scriptSegments[4]
-      )?.toHSL() ?? defaultColor;
-  }
-
-  function changeBeautify() {
-    beautyMode = beautify ? 0 : 1;
-    sendData();
-    dispatch("sync");
-  }
-
-  function sendData() {
-    const script = Script.toScript({
-      short: config.short,
-      array: [...scriptSegments, beautyMode],
-    });
-
-    dispatch("update-action", { short: config.short, script: script });
-  }
-
-  const _suggestions = [
-    [],
-    [
-      { value: "1", info: "layer 1" },
-      { value: "2", info: "layer 2" },
-    ],
-    [],
-    [],
-    [],
-  ];
-
-  let suggestions = [];
-
-  $: if ($event) {
-    updateSuggestions();
-  }
-
-  function updateSuggestions() {
-    const actions = $event.config;
-    const index = actions.findIndex((e) => e.id === config.id);
-    const localDefinitions = LocalDefinitions.getFrom({
-      configs: actions,
-      index: index,
-    });
-    suggestions = _suggestions.map((s, i) => {
-      if (i === 1) {
-        const target = event.parent as GridElement;
-        switch (target.type) {
-          case ElementType.BUTTON:
-            return [
-              { value: "1", info: "Button layer" },
-              { value: "2", info: "Unused layer" },
-            ];
-          case ElementType.ENCODER:
-            return [
-              { value: "1", info: "Button layer" },
-              { value: "2", info: "Rotation layer" },
-            ];
-          case ElementType.FADER:
-            return [
-              { value: "1", info: "Fader layer" },
-              { value: "2", info: "Unused layer" },
-            ];
-          case ElementType.POTMETER:
-            return [
-              { value: "1", info: "Potmeter layer" },
-              { value: "2", info: "Unused layer" },
-            ];
-        }
-      } else {
-        return [...localDefinitions, ...s];
-      }
-    });
-  }
-
-  onMount(() => {
-    updateSuggestions();
-  });
-
-  function updateColor(e: any) {
-    const color = e.detail.color.toRGB();
-    scriptSegments[2] = color.r;
-    scriptSegments[3] = color.g;
-    scriptSegments[4] = color.b;
-
-    sendData();
-    dispatch("sync");
-  }
-
-  enum ColorPickerModel {
-    Square,
-    Slider,
-    Circle,
-  }
-
-  const colorPickerComponent = new Map([
-    [ColorPickerModel.Square, SquareColorPicker],
-    [ColorPickerModel.Slider, SliderColorPicker],
-    [ColorPickerModel.Circle, CircleColorPicker],
-  ]);
-
-  const options = [
-    { title: "Circle", value: ColorPickerModel.Circle },
-    { title: "Square", value: ColorPickerModel.Square },
-    { title: "Slider", value: ColorPickerModel.Slider },
-  ];
-  let selected = ColorPickerModel.Circle;
 </script>
 
-<config-led-color class="flex flex-col gap-4 w-full p-2 pointer-events-auto">
-  <div class="flex flex-row w-full justify-between">
-    <span class="text-white text-sm">Color</span>
-    <div class="flex flex-row gap-1">
-      <div
-        class="flex w-5 h-5 text-center rounded-full border border-black relative bg-secondary"
-      >
-        <span
-          class="absolute inset-0 flex items-center justify-center text-white"
-          >+</span
-        >
-      </div>
-      <div
-        class="flex w-5 h-5 text-center rounded-full border border-black relative bg-secondary"
-      >
-        <span
-          class="absolute inset-0 flex items-center justify-center text-white"
-          >-</span
-        >
-      </div>
-    </div>
-  </div>
-
-  <div class="relative flex w-full h-[15px]">
-    <div class="absolute inset-0 bg-checkboard opacity-20" />
-    <div class="absolute w-full h-full grid grid-cols-1">
-      <div
-        class="flex flex-grow h-full"
-        style="background-image: linear-gradient(to right, transparent , {color.toCSS()});"
-      />
-    </div>
-  </div>
-
-  <div
-    class="grid grid-cols-2 w-full gap-2 min-h-32 text-white place-items-center"
-  >
-    <div class="flex flex-grow bg-red-500 text-white">
-      <MeltSelect bind:target={selected} {options} disabled={false} />
-    </div>
-    <span class="text-white">Intensity</span>
-    <svelte:component
-      this={colorPickerComponent.get(selected)}
-      {color}
-      on:change={updateColor}
-    />
-    <div class="flex h-full w-5 bg-lightness" />
-  </div>
-
-  <SendFeedback
-    feedback_context="LedColor"
-    class="mt-2 text-sm text-gray-500"
-  />
-</config-led-color>
-
-<style>
-  .bg-checkboard {
-    background-image: url('data:image/svg+xml;utf8,<svg xmlns="http://www.w3.org/2000/svg" width="100" height="100" viewBox="0 0 100 100"><rect width="50" height="50" fill="white" /><rect x="50" y="50" width="50" height="50" fill="white" /><rect x="50" width="50" height="50" fill="transparent" /><rect y="50" width="50" height="50" fill="transparent" /></svg>');
-    background-size: 10px 10px;
-    background-repeat: repeat;
-  }
-
-  .bg-lightness {
-    background: linear-gradient(to top, black, white);
-  }
-</style>
+<Temp {config} {index} />
